HU-7 Y HU-9 completas, HU-2 presenta un error

+++ LN/General/Conversiones/Productos/ConvertirAProductosDto/ConvertirAProductosDto.cs
@@ -17,7 +17,7 @@ namespace LN.General.Conversiones.Productos.ConvertirAProductosDto
             {
                 nombre = elProducto.nombre,
                 descripcion = elProducto.descripcion,
-                precio = (float)elProducto.precio,
+                precio = elProducto.precio, 
                 stock = elProducto.stock,
                 imagen = elProducto.imagen,
                 CodigoDelProducto = elProducto.CodigoDelProducto,
